requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.0
groq==0.5.0
PyMuPDF==1.23.21
markdown==3.5.2
